[Add] CameraViewController에서 Camera 기능 분리
